refactor: extract app runtime modes
