+++ crates/mcp-compressor-core/src/app/mod.rs
@@ -1,2 +1,3 @@
 pub mod options;
 pub mod paths;
+pub mod runtime;
